requirements.txt
numpy==1.20.2
torch==1.8.1+cu101
-f https://download.pytorch.org/whl/torch_stable.html
pandas
scikit-learn
ipdb
matplotlib==3.4.2
SQLAlchemy==1.4.1
psycopg2==2.8.6
wget>=3.2
pyunpack>=0.1.2
python-dotenv
# tensorflow==2.5.0rc0